+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+
+* Added the V&K version of Request Foreign Aid; in either case you need to manually enable/disable either activity in the **Manage Activities** button at the top of the **Turn** section
+
 ## [0.44.0] - 2024-05-04
 
 ### Added